Split Up Main
Separated the interactive elements into their own classes to tidy up main as it was getting too large.

Completed the generate grid process by adding in a new message box for the user to select difficulty. Streamlined the message box creation to include a type argument to determine buttons used.